feat: making the nfInstanceId configurable
Initially the NF created at each boot a new uuid. This makes it difficult to retrieve the information afterwards. This commit propose is to add the nfInstanceId in the configuration as optional so that if needed, it can be tracked more easily.
If no nfInstanceId specified, it will still be generated at each boot and give more control to users that might need it (for kube deployment, other checks, better identification)